requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Time Series & Causal Analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Graph algorithms
networkx>=3.1
pygraphviz>=1.11  # Optional: for better graph visualization

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging & utilities
tqdm>=4.65.0
colorlog>=6.7.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Data handling
yfinance>=0.2.18  # For downloading stock data
requests>=2.31.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Optional: Causal discovery libraries
# cdt>=0.6.0  # Causal Discovery Toolbox (optional, has many dependencies)
# causalnex>=0.12.0  # Bayesian Networks library
